+++ aws-sdk-core/lib/aws-sdk-core/emr.rb
@@ -2,4 +2,5 @@ Aws.add_service(:EMR, {
   api: File.join(Aws::API_DIR, 'EMR.api.json'),
   docs: File.join(Aws::API_DIR, 'EMR.docs.json'),
   paginators: File.join(Aws::API_DIR, 'EMR.paginators.json'),
+  waiters: File.join(Aws::API_DIR, 'EMR.waiters.json'),
 })