+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Unreleased Changes
 ------------------
 
+* Feature - Waiters - Added support for version 2 of the waiters
+  definition format.
+
 * Issue - Pagingation - Resolved an issue where an empty hashes or arrays
   were being treated as valid next tokens for paging response. This appears
   to only have affected `Aws::DynamoDB::Client#batch_get_item`.